move rest of common components
