+++ src/App.vue
@@ -63,7 +63,7 @@ import { CHECK_FOR_UPDATES } from '@/messages'
 import TitleBar from '@/common/components/title-bar.vue'
 import Navbar from '@/components/Navbar/Navbar.vue'
 import PlayerContainer from '@/components/Player/Container.vue'
-import ToastOverlay from '@/components/ToastOverlay.vue'
+import ToastOverlay from '@/common/components/toast-overlay.vue'
 import AboutModal from '@/components/Modals/AboutModal.vue'
 import EditModal from '@/components/Modals/EditModal.vue'
 import ManualSearchModal from '@/components/Modals/ManualSearch/ManualSearchModal.vue'
