+++ src/common/components/icon.vue
