+++ src/components/Modals/ManualSearch/CrunchyrollEditor.vue
@@ -49,9 +49,9 @@ import { mdiArrowLeft, mdiCheck } from '@mdi/js'
 import { cacheEpisodes } from '@/common/mutations/episodes'
 import { EpisodeListEpisodes, Provider, Maybe } from '@/graphql/types'
 import Loading from '@/common/components/loading.vue'
-import Icon from '@/components/Icon.vue'
-import CButton from '@/components/CButton.vue'
-import AnimeBanner from '@/components/AnimeBanner.vue'
+import Icon from '@/common/components/icon.vue'
+import CButton from '@/common/components/button.vue'
+import AnimeBanner from '@/common/components/anime-banner.vue'
 import CrunchyrollCollection from './CrunchyrollCollection.vue'
 
 import { Query, Required } from '@/decorators'
