+++ src/components/Player/Player.vue
@@ -113,7 +113,7 @@ import { DISCORD_PAUSE_WATCHING, DISCORD_SET_WATCHING } from '@/messages'
 import { Levels, Stream } from '@/types'
 import { capitalize, clamp, getRelations, lastItem } from '@/utils'
 
-import Icon from '../Icon.vue'
+import Icon from '@/common/components/icon.vue'
 import Controls from './Controls.vue'
 import NextEpisodeOverlay from './NextEpisodeOverlay.vue'
 import EndOfSeasonOverlay from './EndOfSeasonOverlay.vue'
