+++ src/views/settings/settings.vue
@@ -248,9 +248,9 @@ import { mdiInformationOutline, mdiRefresh, mdiUndoVariant } from '@mdi/js'
 
 import LoginHD from '@/common/components/login/hidive.vue'
 import LoginCR from '@/common/components/login/crunchyroll.vue'
-import Checkbox from '@/components/Checkbox.vue'
-import CButton from '@/components/CButton.vue'
-import Icon from '@/components/Icon.vue'
+import Checkbox from '@/common/components/form/checkbox.vue'
+import CButton from '@/common/components/button.vue'
+import Icon from '@/common/components/icon.vue'
 import Dropdown, { DropdownItem } from '@/common/components/form/dropdown.vue'
 import Connection from './components/connection.vue'
 import Keybinding from './components/keybinding.vue'
